+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - BREAKING: fixed typo in the technical accessibility concept scheme identifier
 - BREAKING: fixed copy-pasta in merged-resource-series identifier
 - fixed typos and grammar in descriptions: comma before "that", punctuation, whitespaces
+- fixed issues in entity jsons
 
 ### Security